# openapi.yaml
openapi: 3.0.3
info:
  title: Testimonial Management API
  description: RESTful API for collecting, managing, and analyzing video testimonials
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:3000
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Testimonial:
      type: object
      properties:
        testimonialId:
          type: string
          format: uuid
          example: '550e8400-e29b-41d4-a716-446655440000'
        userId:
          type: number
          example: 1
        customerName:
          type: string
          example: 'John Doe'
        customerEmail:
          type: string
          format: email
          example: 'john@example.com'
        customerPhone:
          type: string
          example: '+1234567890'
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        text:
          type: string
          example: 'Amazing service!'
        videoUrl:
          type: string
          format: uri
          example: 'https://example.com/video.mp4'
        status:
          type: string
          enum: [draft, recording, processing, completed, shared]
          example: 'draft'
        consentGiven:
          type: boolean
          example: true
        sharedAt:
          type: string
          format: date-time
        sharedChannels:
          type: array
          items:
            type: string
            enum: [email, sms, facebook, instagram]
        isDeleted:
          type: boolean
        deletedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /api/testimonials:
    get:
      summary: Get testimonials
      description: Get list of testimonials with pagination and filtering
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, recording, processing, completed, shared]
        - name: sort
          in: query
          schema:
            type: string
            enum: [createdAt, updatedAt, rating]
            default: createdAt
      responses:
        '200':
          description: Testimonials retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  status:
                    type: string
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Testimonial'
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                      page:
                        type: integer
                      limit:
                        type: integer
                      pages:
                        type: integer

    post:
      summary: Create testimonial
      description: Create a new testimonial
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - customerName
              properties:
                customerName:
                  type: string
                  minLength: 2
                  maxLength: 100
                customerEmail:
                  type: string
                  format: email
                customerPhone:
                  type: string
                videoUrl:
                  type: string
                  format: uri
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                text:
                  type: string
                consentGiven:
                  type: boolean
      responses:
        '201':
          description: Testimonial created
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  status:
                    type: string
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Testimonial'
...